move over all learning objects routes
