--- src/app/admin/core/collection.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { CollectionService } from './collection.service';
-
-describe('CollectionService', () => {
-  let service: CollectionService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({ teardown: { destroyAfterEach: false } });
-    service = TestBed.inject(CollectionService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
