+++ src/app/onion/core/learning-object.service.ts
@@ -85,44 +85,6 @@ export class LearningObjectService {
   }
 
   /**
-   * Sends serialized Learning Object to API for creation
-   * Returns new Learningbject's ID on success
-   * Returns error on error
-   *
-   * @param {any} learningObject
-   * @returns {Promise<string>}
-   * @memberof LearningObjectService
-   */
-  create(learningObject): Promise<LearningObject> {
-    const route = LEARNING_OBJECT_ROUTES.CREATE_LEARNING_OBJECT();
-
-    return this.http
-      .post(
-        route,
-        { object: learningObject },
-        { headers: this.headers, withCredentials: true }
-      )
-      .pipe(
-        catchError(this.handleError)
-      )
-      .toPromise()
-      .then((res: any) => {
-        res.id = res._id;
-
-        const learningObject: LearningObject = new LearningObject(res);
-
-        // Fetch the author of the learning object and set it as the
-        // author of the learning object
-        this.userService.getUser(res.authorID).then(user => {
-          learningObject.author = user;
-        });
-
-        return learningObject;
-      });
-    // TODO: Verify this response gives the learning object name
-  }
-
-  /**
    * Creates a Revision of an existing learning object
    *
    * @param learningObjectId
@@ -230,36 +192,6 @@ export class LearningObjectService {
       .then((learningObjects: any[]) => {
         return learningObjects.map(learningObject => new LearningObject(learningObject));
       });
-  }
-
-  /**
-   * Sends updated Learning Object to API for updating.
-   * Returns null success.
-   * Returns error on error
-   *
-   * @param {string} id
-   * @param {LearningObject} learningObject
-   * @returns {Promise<{}>}
-   * @memberof LearningObjectService
-   */
-  // TODO type this parameter
-  save(
-    id: string,
-    learningObject: Partial<LearningObject>,
-    reason?: string,
-  ): Promise<{}> {
-    const route = LEARNING_OBJECT_ROUTES.UPDATE_LEARNING_OBJECT(id);
-    return this.http
-      .patch(
-        route,
-        { updates: learningObject, reason },
-        { headers: this.headers, withCredentials: true, responseType: 'text' }
-      )
-      .pipe(
-
-        catchError(this.handleError)
-      )
-      .toPromise();
   }
 
   /**
@@ -393,83 +325,7 @@ export class LearningObjectService {
     await this.bundlingService.bundleLearningObject(learningObjectId);
   }
 
-  /**
-   * Sends Learning Object's ID to API for deletion
-   *
-   * @param {(string)} id
-   * @returns {Promise<{}>}
-   * @memberof LearningObjectService
-   */
-  delete(learningObjectId: string): Promise<{}> {
-    const route = LEARNING_OBJECT_ROUTES.DELETE_LEARNING_OBJECT(
-      learningObjectId
-    );
-    return this.http
-      .delete(route, {
-        headers: this.headers,
-        withCredentials: true,
-        responseType: 'json'
-      })
-      .pipe(
-        catchError(this.handleError)
-      )
-      .toPromise();
-  }
-
-  /**
-   * Method to delete multiple learning objects
-   *
-   * @param learningObjectIds Array of learning object ids to delete
-   * @returns Promise of all delete requests
-   */
-  deleteMultiple(learningObjectIds: string[]): Promise<any> {
-    const deletePromises = learningObjectIds.map(objectId =>
-      this.http.delete(LEARNING_OBJECT_ROUTES.DELETE_LEARNING_OBJECT(objectId), {
-        headers: this.headers,
-        withCredentials: true,
-        responseType: 'text'
-      })
-      .pipe(
-        catchError(this.handleError)
-      )
-      .toPromise()
-    );
-
-    return Promise.all(deletePromises);
-  }
-
-  setChildren(
-    learningObjectId: string,
-    children: string[],
-    remove: boolean,
-  ): Promise<any> {
-    const removeRoute = LEARNING_OBJECT_ROUTES.REMOVE_LEARNING_OBJECT_CHILD(learningObjectId);
-    const addRoute = LEARNING_OBJECT_ROUTES.UPDATE_LEARNING_OBJECT_CHILDREN(learningObjectId);
-    if (remove) {
-      return this.http
-        .patch(
-          removeRoute,
-          { childObjectId: children[0] },
-          { headers: this.headers, withCredentials: true, responseType: 'text' }
-        )
-        .pipe(
-          catchError(this.handleError)
-        )
-        .toPromise();
-    } else {
-      return this.http
-        .post(
-          addRoute,
-          { childrenIds: children },
-          { headers: this.headers, withCredentials: true, responseType: 'text' }
-        )
-        .pipe(
-
-          catchError(this.handleError)
-        )
-        .toPromise();
-    }
-  }
+  
 
   async updateReadme(id: string): Promise<any> {
     return await this.http.patch(
